Refactor: Move try-it tests to its own dedicated file
